enhancements and include domain
